pyproject.toml
[project]
name = "smartnotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.1",
    "apscheduler>=3.11.0",
    "click>=8.3.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.120.1",
    "numpy>=2.3.4",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "sentence-transformers>=5.1.2",
    "sqlalchemy>=2.0.44",
    "tomli>=2.3.0",
    "uvicorn[standard]>=0.38.0",
    "vadersentiment>=3.3.2",
    "watchdog>=6.0.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "isort>=7.0.0",
    "mypy>=1.18.2",
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "ruff>=0.14.2",
]
llm = [
    "anthropic>=0.72.0",
    "google-generativeai>=0.8.5",
    "openai>=2.6.1",
]

[tool.hatch.build.targets.wheel]
packages = ["smartnotes"]

[project.scripts]
smart = "smartnotes.cli:cli"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.uv]
package = true
